refactored ModalProvider to exist as a global commponent for use with any app modals
