+++ src/global/components/ModalProvider/index.js
@@ -1,0 +1,4 @@
+import ModalProvider from './ModalProvider'
+
+
+export default ModalProvider;
